\name{cleanabs-methods}
\docType{methods}
\alias{cleanabs-methods}
\alias{cleanabs,Abstracts-method}
\title{Methods for Function \code{cleanabs}}
\description{To clean 'NONE' part of searchabsL output.}
\section{Methods}{
\describe{
\item{\code{signature(object = "Abstracts")}}{
From an S4 object of class 'Abstracts' the cleanabs function is able to clean the output of searchabsL by removing the 'NONE' part of resulted abstracts.
}
}}
\keyword{methods}
