\name{combineabs-methods}
\docType{methods}
\alias{combineabs-methods}
\alias{combineabs,Abstracts-method}
\title{\code{Abstracts} Method to Combine Abstracts}
\description{
	\code{combineabs} method to combine the abstracts. object1 and object2 are from \code{Abstracts} class.}
\section{Methods}{
\describe{
\item{\code{signature(object1 = "Abstracts")}}{An S4 object of class "Abstracts"}
\item{\code{signature(object2 = "Abstracts")}}{An S4 object of class "Abstracts"}
}}
\keyword{methods}
