\name{common_words_new}
\alias{common_words_new}
\docType{data}
\title{
R Data containing words which frequently in text
}
\description{
This dataset is used to remove common words from the abstracts. This step is used for size reduction for further data mining.
}
\usage{data(common_words_new)}
\format{
  The format is:
 chr "common_words_new"
}
\details{
The dataset containing common words used to remove them from the text for size reduction.
}
\examples{
data(common_words_new)
}
\keyword{datasets}
