\name{contextSearch-methods}
\docType{methods}
\alias{contextSearch-methods}
\alias{contextSearch,Abstracts-method}
\title{Method for Context Search}
\description{\code{contextSearch} will search the sentence for the given term(s).}
\section{Methods}{
\describe{

\item{\code{signature(object = "Abstracts")}}{
The object from where it will search should be an S4 object of class Abstracts
}
}}
\keyword{methods}
