\name{Find_conclusion}
\alias{Find_conclusion}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
To find the conclusion from the abstract(s).
}
\description{
This function is designed for the user convinience, so that user can get the conclusion from the abstract(s) with out reading the whole abstract(s).
}
\usage{
Find_conclusion(y)
}
\arguments{
  \item{y}{
An S4 object of class 'Abstract'.
}
}
\value{
A list containing conclusion of given abstract(s)
}
\author{
S.Ramachandran, Jyoti Sharma
}
\examples{
\dontrun{res1 = Find_conclusion(y)}
##  here 'y' is an S4 object of class Abstract.
}
\keyword{Function}
