\name{GeneToEntrez}
\alias{GeneToEntrez}
\docType{data}
\title{Data containing Entrez Ids}
\description{This dataset is used in \code{DAVID_info} function of the package, and it contains the Entrez Ids for the respective genes and these Entrez Ids will be used to get information about human genes.}
\usage{data(GeneToEntrez)}
\format{
  The format is:
 chr "GeneToEntrez"
  }
\examples{
data(GeneToEntrez)
}
\keyword{datasets}
