\name{Genewise-methods}
\docType{methods}
\alias{Genewise-methods}
\alias{Genewise,Abstracts-method}
\title{method to find the abstracts for the given gene.}
\description{\code{Genewise} The method Genewise will automatically report the numbers of abstracts for a given gene. It will write the result in the text file named "dataout.txt"}
\section{Methods}{
\describe{

\item{\code{signature(object = "Abstracts")}}{This method will search in an S4 object, containiing abstracts. It will write a text file named "dataout.txt", containing the number of abstracts for the query gene terms} 
}
}
\keyword{methods}

