\name{get_PMCIDS}
\alias{get_PMCIDS}
\title{
To extract the PMC Ids of the abstracts.
}
\description{
\code{get_PMCIDs} is used to fetch the PMC Ids of the abstracts from the corpus.
}
\usage{
get_PMCIDS(abs)
}
\arguments{
  \item{abs}{
\code{abs}an S4 object of class Abtsracts.
}
}
\value{
It returns a list containing PMC Ids.
}
\author{
S.Ramachandran
}
\examples{
\dontrun{get_PMCIDS(abstracts)}
}
\keyword{function}
