\name{getabs-methods}
\docType{methods}
\alias{getabs-methods}
\alias{getabs,Abstracts-method}
\title{\code{getabs} To Get abstracts for a term}
\description{\code{getabs} will search for the abstracts of a given term. It is case sensitive.}
\section{Methods}{
\describe{

\item{\code{signature(object = "Abstracts")}}{This method takes three arguments, first 'object' containing data to be search, 'x', the term to be search, 'y' is logical if set "YES" will consider the case of text.}
}
}
\keyword{methods}
