\name{getabsT-methods}
\docType{methods}
\alias{getabsT-methods}
\alias{getabsT,Abstracts-method}
\title{To Get Abstracts}
\description{
\code{getabsT} will automatically return the abstracts of a term from the data.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Abstracts")}}{
getabsT will search for the abstracts of a term in the data, and will automatically write the number of abstracts into a text file named "dataout.txt".
}
}}
\keyword{methods}

