\name{printabs}
\alias{printabs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{To prind the total number of abstracts in an S4 object of class Abstracts , its start and end}
\description{It gives overview of the abstracts in an S4 object of class Abstracts.}
\usage{
printabs(object)
}
\arguments{
  \item{object}{An S4 object of class Abstracts.}
}
\value{prints the total number of abstracts in an S4 object with additional information.}
\author{S.Ramachandran}
\examples{
\dontrun{printabs(res1)}
## here 'res1' is an S4 object of class Abstracts.
}
\keyword{Function}

