\name{ready}
\alias{ready}
\title{To Initiate the Classes.}
\description{\code{ready} will initiate the classes neccessary for other functions.}
\usage{
ready()
}
\details{This function is neccessary to initiate the classes which are needed for the implementation of other functions. }
\value{ classes}
\author{Dr.S.Ramachandran}
\examples{
\dontrun{ready()}
}
\keyword{function}

