\name{removeabs-methods}
\docType{methods}
\alias{removeabs-methods}
\alias{removeabs,Abstracts-method}
\title{\code{removeabs} To remove abstracts of a term from the data.}
\description{\code{removeabs} This function will search for the abstracts containing the given term to remove them from the data.}
\section{Methods}{
\describe{

\item{\code{signature(object = "Abstracts")}}{This method depicts its function, it will remove the abstracts from the data, and the number of abstracts removed will be written the text file named "dataout.txt"}
}
}
\keyword{methods}

