\name{searchabsT-methods}
\docType{methods}
\alias{searchabsT-methods}
\alias{searchabsT,Abstracts-method}
\title{ \code{searchabsT} Searching abstracts}
\description{
\code{searchabsT} will perform a specific search for the given term.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Abstracts")}}{
It is similar to the searchabsL method, but it is more specific than searchabsL, it is case sensitive, however searchabsL is not. 
}
}}
\keyword{methods}

