\name{sendabs-methods}
\docType{methods}
\alias{sendabs-methods}
\alias{sendabs,Abstracts-method}
\title{To send the Data into a File}
\description{
\code{sendabs} will write the data of an object of class 'Abstracts' into a tab delimited text file with header Journal, Abstract, and  
PMID}
\section{Methods}{
\describe{
\item{\code{signature(object = "Abstracts")}}{
sendabs will send the data into a text file. It writes a tab delimited text file for PubMed abstracts containing Journal, Abstract, and PMID.
}
}}
\keyword{methods}

