\name{subabs-methods}
\docType{methods}
\alias{subabs-methods}
\alias{subabs,Abstracts-method}
\title{Getting subabstracts}
\description{
\code{subabs} subabs will extract the sub abstracts corresponding to a given range, from the whole data.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Abstracts")}}{
From an S4 object of class 'Abstracts' the subabs function is able to extract the abstracts corresponding to a given range.
}
}}
\keyword{methods}

