\name{subsetabs-methods}
\docType{methods}
\alias{subsetabs-methods}
\alias{subsetabs,Abstracts-method}
\title{To make subset of Abstracts.}
\description{
\code{subsetabs} is used to subset of Abstracts from the large corpus. Its output is used in other functions like currentabs_fn and previousabs_fn
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Abstracts")}}{
subsetabs will divide the large corpus into subset. 
}
}}
\keyword{methods}

