\name{Yearwise-methods}
\docType{methods}
\alias{Yearwise-methods}
\alias{Yearwise,Abstracts-method}
\title{ \code{Yearwise} Year wise extraction of Abstracts }
\description{ \code{Yearwise} will report the abstracts for given year(s). }
\section{Methods}{
\describe{

\item{\code{signature(object = "Abstracts")}}{ This method "Yearwise" is written to fetch the abstracts yearly. }
}}
\keyword{methods}

