\name{HGNC2UniprotID}
\alias{HGNC2UniprotID}
\docType{data}
\title{ R Data containing HGNC2UniprotID data mapping. }
\description{ This dataset contains HGNC2UniprotID from Uniprot and is used in uniprotfn() function of this package, to get the information of a gene from the Uniprot. }
\usage{data(HGNC2UniprotID)}
\format{
  The format is:
 chr "HGNC2UniprotID"
}
\details{ The dataset contains HGNC2UniprotID }
\examples{
data(HGNC2UniprotID)
}
\keyword{datasets}
