\name{HGNCdata}
\alias{HGNCdata}
\docType{data}
\title{ R Data containing HGNC data. }
\description{ This dataset contains data from Human Gene Nomenclature Committe i.e HGNC ID, HGNC approved symbol, approved name, gene synonyms, chromosome no., accession numbers and RefSeq ids.}
\usage{data(HGNCdata)}
\format{
  The format is:
 chr "HGNCdata"
}
\details{ The dataset contains HGNCdata }
\examples{
data(HGNCdata)
}
\keyword{datasets}
